tasks handler fixed

+++ backend/tasks/delete_handler.py
@@ -1,9 +1,30 @@
 from collections import defaultdict
 from urllib.parse import urlparse
 from google.cloud import storage
-import logging
+import logging.config
 
-logging.basicConfig(level=logging.INFO)
+LOGGING = {
+    "version": 1,
+    "disable_existing_loggers": False,
+    "formatters": {
+        "google_json_formatter": {
+            "()": "pythonjsonlogger.jsonlogger.JsonFormatter",
+            "format": "%(message)s",
+        }
+    },
+    "handlers": {
+        "google_cloud_handler": {
+            "class": "google.cloud.logging.handlers.StructuredLogHandler",
+            "formatter": "google_json_formatter"
+        },
+    },
+    "root": {
+        "handlers": ["google_cloud_handler"],
+        "level": "INFO",
+    }
+}
+
+logging.config.dictConfig(LOGGING)
 storage_client = storage.Client()
 
 def image_delete(data):
@@ -33,13 +54,22 @@ def image_delete(data):
     for bucket_name, blob_names in blobs_by_bucket.items():
         try:
             bucket = storage_client.bucket(bucket_name)
-            batch_size = 100
+            
+            prefix = f"users/{blob_names[0].split('/')[1]}/"
+            existing_blobs = {blob.name for blob in bucket.list_blobs(prefix=prefix)}
+            blobs_to_delete = [name for name in blob_names if name in existing_blobs]
 
-            for i in range(0, len(blob_names), batch_size):
+            if not blobs_to_delete:
+                continue
+
+            batch_size = 100
+            for i in range(0, len(blobs_to_delete), batch_size):
                 with storage_client.batch():
-                    batch = blob_names[i:i + batch_size]
+                    batch = blobs_to_delete[i:i + batch_size]
                     for blob_name in batch:
                         blob = bucket.blob(blob_name)
                         blob.delete()
+            
+            logging.info(f"Successfully requested deletion", extra={'bucket_name': bucket_name, 'blobs_to_delete': blobs_to_delete})
         except Exception as e:
-            logging.error("Failed to process batch deletion for bucket", extra={'bucket_name': bucket_name, 'blob_names': blob_names, 'error': str(e)}, exc_info=True)
+            logging.error("Failed to process batch deletion for bucket", extra={'bucket_name': bucket_name, 'error': str(e)}, exc_info=True)
